requirements.txt
aiosignal==1.4.0
aiosqlite==0.19.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
black==23.12.1
certifi==2026.2.25
cffi==2.0.0
cryptography==46.0.3
click==8.3.1
colorama==0.4.6
distro==1.9.0
ecdsa==0.19.1
et_xmlfile==2.0.0
fastapi==0.135.1
filelock==3.25.1
frozenlist==1.8.0
fsspec==2026.2.0
greenlet==3.3.2
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httptools==0.7.1
httpx==0.26.0
idna==3.11
iniconfig==2.3.0
jiter==0.13.0
loguru==0.7.2
lxml==6.0.2
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==1.24.3
openai==2.26.0
openpyxl==3.1.2
packaging==26.0
pandas==2.0.3
pathspec==1.0.4
pillow==10.2.0
platformdirs==4.9.4
pluggy==1.6.0
propcache==0.4.1
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.1.0
pydantic_core==2.41.5
Pygments==2.19.2
pytest==7.4.4
pytest-asyncio==0.23.8
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.2.2
python-multipart==0.0.6
pywebview==5.4
pytz==2026.1.post1
PyYAML==6.0.1
rich==14.3.3
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.25
starlette==0.52.1
tqdm==4.66.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.27.0
watchfiles==1.1.1
websockets==12.0
win32_setctime==1.2.0
yarl==1.23.0
